flask-jwt/jwt-api/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
PyJWT==2.8.0